Switch to different binary targets
