+++ oracle-ingestor-lambda-parquet/Cargo.toml
@@ -2,9 +2,15 @@
 name = "oracle-ingestor-lambda"
 version = "0.1.0"
 edition = "2021"
+default-run = "oracle-ingestor-lambda"
 
-[features]
-local = []
+[[bin]]
+name = "oracle-ingestor-lambda"
+path = "src/main.rs"
+
+[[bin]]
+name = "local"
+path = "src/local.rs"
 
 [dependencies]
 anyhow = "1"
@@ -13,7 +19,8 @@ helium-crypto = {version = "0.6.3"}
 helium-proto = {git = "https://github.com/helium/proto", branch = "master", features = ["services"]}
 file-store = {git = "https://github.com/helium/oracles", branch = "main"}
 futures = "*"
-lambda_runtime = "0.7"
+lambda_runtime = "0.8.0"
+lambda_http = "0.8.0"
 serde =  {version = "1", features=["derive"]}
 serde_json = "1"
 sqlx = { version = "0.6", features = [ "runtime-tokio-rustls" , "postgres", "chrono", "migrate"] }
