+++ oracle-ingestor-lambda-parquet/src/main.rs
@@ -1,58 +1,23 @@
-use anyhow::anyhow;
-use anyhow::Result;
-use clap::Parser;
-use oracle_ingestor_lambda::{
-    cli::{current, history},
-    settings::Settings,
-};
-use serde::Serialize;
-use std::path;
+use lambda_runtime::{run, service_fn, Error, LambdaEvent};
+use serde_json::Value;
 
-#[derive(Debug, Serialize)]
-struct FailureResponse {
-    pub body: String,
-}
-
-#[derive(Debug, clap::Subcommand)]
-pub enum Cmd {
-    /// Run in historical data gathering mode
-    History(history::Cmd),
-    /// Start from current given timestamp
-    Current(current::Cmd),
-}
-
-impl Cmd {
-    pub async fn run(self, settings: Settings) -> Result<()> {
-        match self {
-            Self::History(cmd) => cmd.run(&settings).await,
-            Self::Current(cmd) => cmd.run(&settings).await.map_err(|e| anyhow!(e)),
-        }
-    }
-}
-
-#[derive(Debug, clap::Parser)]
-#[clap(version = env!("CARGO_PKG_VERSION"))]
-#[clap(about = "Oracles Parquet Parser")]
-pub struct Cli {
-    /// Optional configuration file to use. If present the toml file at the
-    /// given path will be loaded. Environment variables can override the
-    /// settings in the given file.
-    #[clap(short = 'c')]
-    config: Option<path::PathBuf>,
-
-    #[clap(subcommand)]
-    cmd: Cmd,
-}
-
-impl Cli {
-    pub async fn run(self) -> Result<()> {
-        let settings = Settings::new(self.config)?;
-        self.cmd.run(settings).await
-    }
+async fn function_handler(event: LambdaEvent<Value>) -> Result<(), Error> {
+    tracing::info!("event: {:?}", event);
+    // TODO
+    // - Get the settings and create handler
+    // - Invoke handler::run with current mode
+    Ok(())
 }
 
 #[tokio::main]
-async fn main() -> Result<()> {
-    let cli = Cli::parse();
-    cli.run().await
+async fn main() -> Result<(), Error> {
+    tracing_subscriber::fmt()
+        .with_max_level(tracing::Level::INFO)
+        // disable printing the name of the module in every log line.
+        .with_target(false)
+        // disabling time is handy because CloudWatch will add the ingestion time.
+        .without_time()
+        .init();
+
+    run(service_fn(function_handler)).await
 }
